Detection & SIEM Projects/Detection-as-Code Pipeline (Sigma + SIEM)/requirements.txt
pyyaml>=6.0
flask>=3.0
requests>=2.31
